MDB-21297: forbit usage of COPY TO PROGRAMM and COPY FROM PROGRAMM to non-superuser
It is well known that some of PostgreSQL-related security issues (CVE)
was related to COPY FROM/TO PROGRAM exploits for priviledge escalation
or other unwanted behaviour or consequences. Thus, proper usage of
this feature needed. For now, simply forbit this.

Allow to COPY to/from file in Cloud. Without this regression tests fails